Rental order feature initiated

+++ app/src/main/java/com/krm/rentalservices/model/Product.kt
@@ -5,7 +5,7 @@ import com.google.firebase.Timestamp
 data class Product(
     var id: String = "",
     var name: String = "",
-    var rentalPrice: Double = 0.0,
+    var rentalPrice: Long = 0,
     var description: String = "",
     var timestamp: Timestamp?
 )
